+++ samples/Basic/Models/InterfaceA.cs
@@ -1,0 +1,12 @@
+﻿namespace Basic.Models
+{
+	public interface IInterfaceA
+	{
+		int Id { get; set; }
+	}
+
+	public class InterfaceAImpl : IInterfaceA
+	{
+		public int Id { get; set; }
+	}
+}